support set token secret through nodePublishSecretRef

+++ pkg/mounter/proxy/server/ossfs/driver.go
@@ -11,6 +11,7 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/mounter"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/mounter/proxy"
 	"github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/mounter/proxy/server"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -21,6 +22,10 @@ import (
 func init() {
 	server.RegisterDriver(NewDriver())
 }
+
+const (
+	OssfsPasswdFile = "passwd-ossfs"
+)
 
 type Driver struct {
 	pids *sync.Map
@@ -47,19 +52,16 @@ func (h *Driver) Mount(ctx context.Context, req *proxy.MountRequest) error {
 	options := req.Options
 
 	// prepare passwd file
-	var passwdFile string
-	if passwd := req.Secrets["passwd-ossfs"]; passwd != "" {
-		tmpDir, err := os.MkdirTemp("", "ossfs-")
-		if err != nil {
-			return err
-		}
-		passwdFile = filepath.Join(tmpDir, "passwd")
-		err = os.WriteFile(passwdFile, []byte(passwd), 0o600)
-		if err != nil {
-			return err
-		}
+	passwdFile, tokenDir, credOpts, err := prepareCredentialFiles(req.Secrets)
+	if err != nil {
+		return err
+	}
+	options = append(options, credOpts...)
+	if passwdFile != "" {
 		klog.V(4).InfoS("created ossfs passwd file", "path", passwdFile)
-		options = append(options, "passwd_file="+passwdFile)
+	}
+	if tokenDir != "" {
+		klog.V(4).InfoS("created ossfs token directory", "dir", tokenDir)
 	}
 
 	args := mount.MakeMountArgs(req.Source, req.Target, "", options)
@@ -70,7 +72,7 @@ func (h *Driver) Mount(ctx context.Context, req *proxy.MountRequest) error {
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
-	err := cmd.Start()
+	err = cmd.Start()
 	if err != nil {
 		return fmt.Errorf("start ossfs failed: %w", err)
 	}
@@ -93,9 +95,17 @@ func (h *Driver) Mount(ctx context.Context, req *proxy.MountRequest) error {
 			klog.InfoS("ossfs exited", "mountpoint", target, "pid", pid)
 		}
 		ossfsExited <- err
-		if err := os.Remove(passwdFile); err != nil {
-			klog.ErrorS(err, "Remove passwd file", "mountpoint", target, "path", passwdFile)
+		if passwdFile != "" {
+			if err := os.Remove(passwdFile); err != nil {
+				klog.ErrorS(err, "Remove passwd file", "mountpoint", target, "path", passwdFile)
+			}
 		}
+		if tokenDir != "" {
+			if err := os.RemoveAll(tokenDir); err != nil {
+				klog.ErrorS(err, "Remove token directory", "mountpoint", target, "dir", tokenDir)
+			}
+		}
+
 		close(ossfsExited)
 	}()
 
@@ -158,4 +168,60 @@ func (h *Driver) Terminate() {
 	// wait all ossfs processes to exit
 	h.wg.Wait()
 	klog.InfoS("All ossfs processes exited")
+}
+
+func writeFile(dir, fileName, contents string, perm os.FileMode) error {
+	file := filepath.Join(dir, fileName)
+	return os.WriteFile(file, []byte(contents), perm)
+}
+
+// prepareCredentialFiles returns:
+//  1. file:    path of ossfs credential file for fixed AKSK
+//  2. dir:     dorectory of ossfs credential files for token
+//  3. options: extra options
+//  4. error
+func prepareCredentialFiles(secrets map[string]string) (file, dir string, options []string, err error) {
+	var tmpDir string
+	tmpDir, err = os.MkdirTemp("", "ossfs-")
+	if err != nil {
+		return
+	}
+
+	// fixed AKSK
+	if passwd := secrets[OssfsPasswdFile]; passwd != "" {
+		err = writeFile(tmpDir, "passwd", passwd, 0o600)
+		if err != nil {
+			return
+		}
+		file = filepath.Join(tmpDir, "passwd")
+		options = append(options, "passwd_file="+file)
+		return
+	}
+
+	// token
+	tokenKey := []string{mounter.KeyAccessKeyId, mounter.KeyAccessKeySecret, mounter.KeySecurityToken, mounter.KeyExpiration}
+	tokenDir := filepath.Join(tmpDir, "token")
+	var token bool
+	for _, key := range tokenKey {
+		val := secrets[filepath.Join(OssfsPasswdFile, key)]
+		if val == "" {
+			continue
+		}
+		err = os.MkdirAll(tokenDir, 0o644)
+		if err != nil {
+			klog.Errorf("mkdirall tokendir failed %v", err)
+			return
+		}
+		err = writeFile(tokenDir, key, val, 0o600)
+		if err != nil {
+			klog.Errorf("writeFile %s failed %v", key, err)
+			return
+		}
+		token = true
+	}
+	if token {
+		dir = tokenDir
+		options = append(options, "passwd_file="+tokenDir)
+	}
+	return
 }
